chore: refactor to account.ts
