+++ packages/better-auth/src/api/routes/index.ts
@@ -9,5 +9,3 @@ export * from "./error";
 export * from "./ok";
 export * from "./sign-up";
 export * from "./account";
-export * from "./refresh-token";
-export * from "./get-access-token";
